refact : atomic design
