+++ src/app/App.js
@@ -1,7 +1,5 @@
 import Home from 'pages/Home';
 
-function App() {
+export default function App() {
     return <Home />;
 }
-
-export default App;
