--- src/components/Buttons/index.js
@@ -1,8 +1,0 @@
-import BlockButton from 'components/Buttons/BlockButton';
-import ChecklistButton from 'components/Buttons/ChecklistButton';
-import EmojiButton from 'components/Buttons/EmojiButton';
-import ImageButton from 'components/Buttons/ImageButton';
-import LinkButton from 'components/Buttons/LinkButton';
-import MarkButton from 'components/Buttons/MarkButton';
-
-export { BlockButton, ChecklistButton, EmojiButton, ImageButton, LinkButton, MarkButton };
