+++ src/components/molecules/EmojiButton.js
@@ -3,10 +3,11 @@ import { useEffect, useRef, useState } from 'react';
 import data from '@emoji-mart/data';
 import Picker from '@emoji-mart/react';
 import { faIcons } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { useSlate } from 'slate-react';
 
 import CustomEditor from 'lib/slate/helpers/customEditor';
+
+import Button from 'components/atoms/Button';
 
 export default function EmojiButton() {
     const editor = useSlate();
@@ -27,14 +28,12 @@ export default function EmojiButton() {
 
     return (
         <div className="relative">
-            <button
-                type="button"
+            <Button
+                icon={faIcons}
                 title="emoji button"
                 onClick={() => setShowEmoji((prev) => !prev)}
                 className={`${showEmoji ? 'bg-yellow-300' : ''} block w-[40px] h-[30px] [line-height:30px] border border-black font-bold uppercase`}
-            >
-                <FontAwesomeIcon icon={faIcons} />
-            </button>
+            />
             {showEmoji && (
                 <div ref={wrapperRef} className="absolute left-[-176px] z-[1030]">
                     <Picker data={data} onEmojiSelect={(e) => CustomEditor.toggleEmoji(editor, e.native)} />
