+++ src/components/molecules/ImageButton.js
@@ -1,10 +1,11 @@
 import { useRef, useState } from 'react';
 
 import { faImage } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { useSlateStatic } from 'slate-react';
 
 import CustomEditor from 'lib/slate/helpers/customEditor';
+
+import Button from 'components/atoms/Button';
 
 export default function ImageButton() {
     const editor = useSlateStatic();
@@ -26,18 +27,16 @@ export default function ImageButton() {
 
     return (
         <>
-            <button
-                type="button"
+            <Button
+                icon={faImage}
                 title="image button"
-                className={`${isActive ? 'bg-black text-white' : ''} block w-[40px] h-[30px] [line-height:30px] border border-black font-bold uppercase`}
                 onMouseDown={(event) => {
                     event.preventDefault();
                     inputRef.current.click();
                     setIsActive(true);
                 }}
-            >
-                <FontAwesomeIcon icon={faImage} />
-            </button>
+                className={`${isActive ? 'bg-black text-white' : ''} block w-[40px] h-[30px] [line-height:30px] border border-black font-bold uppercase`}
+            />
             <input
                 ref={inputRef}
                 id="image"
